% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALA_downloader.R
\name{ALA_downloader}
\alias{ALA_downloader}
\title{Downloads occurrence data from the Atlas of Living Australia (ALA)}
\usage{
ALA_downloader(path, userEmail = NULL, ALA_taxon, DL_reason = 4)
}
\arguments{
\item{path}{A character directory. The path to a folder where the download will be stored.}

\item{userEmail}{A character string. The email to use with your ALA account (you must create and account).}

\item{ALA_taxon}{A character string. The taxon to download from ALA. Uses \code{\link[galah:galah_identify]{galah::galah_identify()}}}

\item{DL_reason}{Numeric. The reason for data download according to \code{\link[galah:galah_config]{galah::galah_config()}}}
}
\value{
Completes an ALA data download and saves those data to the path provided.
}
\description{
Downloads ALA data and creates a new file in the path to put those data
}
\examples{
\dontrun{
remotes::install_github("AtlasOfLivingAustralia/galah@dev")
ALA_downloader(path = DataPath,
               userEmail = "InsertYourEmail",
               ALA_taxon = "Apiformes")
               }
}
