% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeeDCQuery.R
\name{BeeDCQuery}
\alias{BeeDCQuery}
\title{Query the bee taxonomy and country checklist}
\usage{
BeeDCQuery(beeName = NULL, searchChecklist = TRUE, printAllSynonyms = FALSE)
}
\arguments{
\item{beeName}{Character or character vector. A single or several bee species names to search for
in the beesTaxonomy and beesChecklist tables.}

\item{searchChecklist}{Logical. If TRUE (default), search the country checklist for each species.}

\item{printAllSynonyms}{Logical. If TRUE, all synonyms will be printed out for each entered name.
default = FALSE.}
}
\value{
Returns a list with the elements 'taxonomyReport' and 'SynonymReport'. IF searchChecklist
is TRUE, then 'checklistReport' will also be returned.
}
\description{
A simple function to return information about a particular species, including name validity and
country occurrences.
}
\examples{

  # Single entry example
testQuery <- BeeDCQuery(
  beeName = "Homalictus fijiensis",
  searchChecklist = TRUE,
  printAllSynonyms = TRUE)

  # Multiple entry example
testQuery <- BeeDCQuery(
  beeName = c("Homalictus fijiensis", "Homalictus urbanus",
  "Lasioglossum fijiense (Perkins and Cheesman, 1928)"),
  searchChecklist = TRUE,
  printAllSynonyms = TRUE)
  
    # Example way to examine a report from the output list
  testQuery$checklistReport



}
