% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beesChecklist.R
\docType{data}
\name{beesChecklist}
\alias{beesChecklist}
\title{A country-level checklist of bees from Discover Life}
\format{
An object of class \code{"tibble"}
\describe{
\item{validName}{The valid scientificName as it should occur in the scientificName column.}
\item{DiscoverLife_name}{The full country name as it occurs on Discover Life.}
\item{rNaturalEarth_name}{Country name from rnaturalearth's name_long.}
\item{shortName}{A short version of the country name.}
\item{DiscoverLife_ISO}{The ISO country name as it occurs on Discover Life.}
\item{Alpha-2}{Alpha-2 from rnaturalearth.}
\item{Alpha-3}{Alpha-3 from rnaturalearth.}
\item{official}{Official country name = "yes" or only a Discover Life name = "no".}
\item{Source}{A text strign denoting the source or author of the name-country pair.}
\item{matchCertainty}{Quality of the name's match to the Discover Life checklist.}
\item{canonical}{The valid species name without scientificNameAuthority.}
\item{canonical_withFlags}{The validName without the scientificNameAuthority but with Discover Life flags.}
\item{family}{Bee family.}
\item{subfamily}{Bee subfamily.}
\item{genus}{Bee genus.}
\item{subgenus}{Bee subgenus.}
\item{infraspecies}{Bee infraSpecificEpithet.}
\item{infraspecificEpithet}{Bee infraspecificEpithet.}
\item{species}{Bee specificEpithet.}
\item{scientificNameAuthorship}{Bee scientificNameAuthorship.}
\item{taxon_rank}{Rank of the taxon name.}
\item{Notes}{Discover Life country name notes.}

}
}
\usage{
data(beesChecklist)
}
\description{
The table contains taxonomic and country information for the bees of the world based on
data collated on Discover Life. See \code{\link[=beesTaxonomy]{beesTaxonomy()}} for further context.
}
\examples{

## Not run:
data(beesChecklist)
head(beesChecklist)
## End(Not run)
}
\references{
This dataset was created using the Discover Life checklist and taxonomy.
for further information on download date in file name.
Dataset is from the publication:
DOREY, J. B., CHESSHIRE, P. R., BOLAÑOS, A. N., O’REILLY, R. L., BOSSERT, S., COLLINS, S. M., LICHTENBERG, E. M., TUCKER, E., SMITH-PARDO, A., FALCON-BRINDIS, A., GUEVARA, D. A., RIBEIRO, B. R., DE PEDRO, D., FISCHER, E., HUNG, J. K.-L., PARYS, K. A., ROGAN, M. S., MINCKLEY, R. L., VELZCO, S. J. E., GRISWOLD, T., ZARRILLO, T. A., SICA, Y., ORR, M. C., GUZMAN, L. M., ASCHER, J., HUGHES, A. C. & COBB, N. S. In review. A globally synthesised and flagged bee occurrence dataset and cleaning workflow. Scientific Data.
}
\keyword{datasets}
