% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beesCountrySubset.R
\docType{data}
\name{beesCountrySubset}
\alias{beesCountrySubset}
\title{A simple scientificName and country_suggested test dataset}
\format{
An object of class \code{"tibble"}
\describe{
\item{scientificName}{Full scientificName as shown on DiscoverLife}
\item{country_suggested}{A country name suggested by the \code{\link[bdc:bdc_country_standardized]{bdc::bdc_country_standardized()}} function.}
}
}
\usage{
data("beesCountrySubset", package = "BeeBDC")
}
\description{
A small bee occurrence dataset with 1,488 scientificName-country_suggested combinations across
four countries; Fiji, Uganda, Vietnam, and Zambia.
}
\examples{

beesCountrySubset <- BeeBDC::beesCountrySubset
head(beesCountrySubset)

}
\references{
This data set was created by during the writing of the paper:
Dorey, J.B., (upcoming) How many bee species are there? A quantitative global estimate. Journal
}
\keyword{datasets}
