% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalData_readRs.R
\name{CAES_readr}
\alias{CAES_readr}
\alias{Col_readr}
\alias{Arm_readr}
\title{Reads specific data files into Darwin Core format}
\usage{
CAES_readr(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = "Sheet1"
)

Col_readr(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = sheet
)

Arm_readr(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = "Sheet1"
)
}
\arguments{
\item{path}{The path to the directory containing the data.}

\item{inFile}{The name of the file itself (can also be the remainder of a path including the file name).}

\item{outFile}{The name of the Darwin Core format file to produce.}

\item{dataLicense}{The license to accompany each record in the Darwin Core 'license' column.}

\item{sheet}{A character String. For those datasets read from an .xlsx format, provide the
sheet name.}
}
\value{
A data frame that's in Darwin Core format.
}
\description{
Data files are specific to various data providers and users may examine code function-by-function.
}
\section{Functions}{
\itemize{
\item \code{Col_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{Arm_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

}}
\examples{
An example using a .csv file
EPEL_Data <- EPEL_readr(path = paste0(DataPath, "/Additional_Datasets"),
inFile = "/InputDatasets/bee_data_canada.csv",
outFile = "jbd_EPEL_data.csv",
dataLicense = "https://creativecommons.org/licenses/by-nc-sa/4.0/")

An example using a .xlsx file
    Arm_Data <- Arm_readr(path = paste0(DataPath, "/Additional_Datasets"),
inFile = "/InputDatasets/Bee database Armando_Final.xlsx",
outFile = "jbd_Arm_Data.csv",
sheet = "Sheet1",
dataLicense = "https://creativecommons.org/licenses/by-nc-sa/4.0/")
}
