% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryNameCleanR.R
\name{countryNameCleanR}
\alias{countryNameCleanR}
\title{A simple function to fix a user-input list of country name issues.}
\usage{
countryNameCleanR(data = NULL, ISO2_table = NULL, commonProblems = NULL)
}
\arguments{
\item{data}{A data frame or tibble. Occurrence records as input.}

\item{ISO2_table}{A data frame or tibble with the columns ISO2 and FullName for country names. Default
is a static version from Wikipedia.}

\item{commonProblems}{A data frame or tibble with the columns problem and fix.}
}
\value{
Returns the input data but with the countries occurring in the problem column replaced
with those in the fix column.
}
\description{
This function is basic for a user to manually fix some country name inconsistencies.
}
\examples{
beesFlagged_out <- countryNameCleanR(
data = BeeDC::beesFlagged,
commonProblems = commonProblems)
}
