% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{data_saver}
\alias{data_saver}
\title{Simple function to save occurrence AND EML data as a list}
\usage{
data_saver(
  path = NULL,
  save_type = NULL,
  occurrences = NULL,
  eml_files = NULL,
  file_prefix = NULL
)
}
\arguments{
\item{path}{The main path to look for data in.}

\item{save_type}{Character. File format to save as. Either "R_file" OR "CSV_file".}

\item{occurrences}{The occurrences to save as a data frame or tibble.}

\item{eml_files}{A list of the EML files.}

\item{file_prefix}{Character. A prefix for the resulting output file.}
}
\value{
Saves both occurrence and EML data as a list.
}
\description{
Generally used at the end of 1.x in the example workflow.
}
\examples{
\dontrun{
data_saver(path = DataPath,# The main path to look for data in
save_type = "CSV_file", # "R_file" OR "CSV_file"
occurrences = Complete_data$Data_WebDL, # The existing datasheet
eml_files = Complete_data$eml_files, # The existing EML files
file_prefix = "Fin_") # The prefix for the filenames
}

}
