% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFindR.R
\name{dateFindR}
\alias{dateFindR}
\title{Find dates in other columns}
\usage{
dateFindR(data = NULL, maxYear = lubridate::year(Sys.Date()), minYear = 1700)
}
\arguments{
\item{data}{A data frame or tibble. Occurrence records as input.}

\item{maxYear}{Numeric. The maximum year considered reasonable to find. Default = lubridate::year(Sys.Date()).}

\item{minYear}{Numeric. The minimum year considered reasonable to find. Default = 1700.}
}
\value{
The input data with updated eventDate, year, month, and day columns.
}
\description{
A function made to search other columns for dates and add them to the eventDate column.
Searches the columns: locality, fieldNotes, locationRemarks, and verbatimEventDate.
}
\examples{
# Run the example dataset... depending on the version you may not find any missing eventDates are rescued.
beesRaw_out <- dateFindR(data = beesRaw,
                         # Years above this are removed (from the recovered dates only)
                         maxYear = lubridate::year(Sys.Date()),
                         # Years below this are removed (from the recovered dates only)
                         minYear = 1700)
}
