% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonAlley.R
\name{diagonAlley}
\alias{diagonAlley}
\title{Find fill-down errors}
\usage{
diagonAlley(data = NULL, minRepeats = NULL)
}
\arguments{
\item{data}{A data frame or tibble. Occurrence records as input.}

\item{minRepeats}{Numeric. The minimum number of lat or lon repeats needed to flag a record}
}
\value{
The function returns the input data with a new column, .sequential, where FALSE =
records that have consecutive latitudes or longitudes greater than or equal to the user-defined
threshold.
}
\description{
A simple function that looks for potential latitude and longitude fill-down errors by
identifying identical coordinates in a user-defined number of consecutive records.
}
\examples{
# Read in the example data
  data(beesRaw)
 # Run the function
  beesRaw_out <- diagonAlley(
    data = beesRaw,
    # The minimum number of repeats needed to find a sequence in for flagging
    minRepeats = 4)
  

}
