% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirMaker.R
\name{dirMaker}
\alias{dirMaker}
\title{Set up global directory paths and create folders}
\usage{
dirMaker(
  RootPath = RootPath,
  ScriptPath = NULL,
  DataPath = NULL,
  DataSubPath = "/Data_acquisition_workflow",
  DiscLifePath = NULL,
  OutPath = NULL,
  OutPathName = "Output",
  RDoc = NULL
)
}
\arguments{
\item{RootPath}{character string. The \code{RootPath} is the path at the root of your project and in
which all other paths should ideally
be located. However, users may specify paths not contained in the \code{Rootpath.}}

\item{ScriptPath}{character string. The \code{ScriptPath} is the path to any additional functions that
you would like to read in.}

\item{DataPath}{character string. The path to the folder that contains occurrence data.}

\item{DataSubPath}{character string. If a \code{DataPath} is not provided, this will be used as the \code{DataPath}
folder name within the \code{RootPath.} Default is "/Data_acquisition_workflow"}

\item{DiscLifePath}{character string. The path to the folder that contains data from Ascher
and Pcikering's Discover Life website.}

\item{OutPath}{character string. The path to the folder where data will be saved.}

\item{OutPathName}{character string. The name of the \code{OutPath} subfolder within the \code{RootPath.} Default is
"Output".}

\item{RDoc}{character string. The path to the current script or report, relative to the project
root. Passing an absolute path raises an error. This argument is used by \code{\link[here:i_am]{here::i_am()}} and incorrectly
setting this may result in bdc:: figures being saved to your computer's root directory}
}
\value{
A list containing the required directories in your global environment. This function should
be run at the start of each session. This function will also create the requested folders,
if they do not exist.
}
\description{
If those folders do not exist, they will be created for you.
}
\examples{
\dontrun{
# Standard/basic usage:
RootPath <- "/Users/jamesdorey/Desktop/Uni/My_papers/Bee_SDM_paper"
dirMaker(
RootPath = RootPath,
# Input the location of the workflow script RELATIVE to the RootPath
RDoc = "BDC_repo/BeeCleaning_SciData.R") \%>\%
  # Add paths created by this function to the .GlobalEnv
  list2env(envir = .GlobalEnv)  
# Set the working directory
setwd(DataPath)

# Custom OutPathName provided
RootPath <- "/Users/jamesdorey/Desktop/Uni/My_papers/Bee_SDM_paper"
  dirMaker(
 RootPath = RootPath,
 # Set some custom OutPath info
 OutPath = NULL,
 OutPathName = "T2T_Output",
 # Input the location of the workflow script RELATIVE to the RootPath
 RDoc = "BDC_repo/BeeCleaning_SciData.R") \%>\%
   # Add paths created by this function to the .GlobalEnv
   list2env(envir = .GlobalEnv)  
 # Set the working directory
 setwd(DataPath)

# Further customisations are also possible
RootPath <- "/Users/jamesdorey/Desktop/Uni/My_papers/Bee_SDM_paper"
dirMaker(
  RootPath = RootPath,
  ScriptPath = "...path/Bee_SDM_paper/BDC_repo/BeeDC/R",
  DiscLifePath = "...path/BDC_repo/DiscoverLife_Data",
  OutPathName = "AsianPerspective_Output",
  # Input the location of the workflow script RELATIVE to the RootPath
  RDoc = "AsianPerspecitve_workflow.R") \%>\%
  # Add paths created by this function to the .GlobalEnv
  list2env(envir = .GlobalEnv)  
# Set the working directory
setwd(DataPath)
}

}
