% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalData_readRs.R
\name{EPEL_readr}
\alias{EPEL_readr}
\alias{ASP_readr}
\alias{BMin_readr}
\alias{BMont_readr}
\alias{Ecd_readr}
\alias{Gai_readr}
\alias{KP_readr}
\alias{INHS_readr}
\alias{EcoS_readr}
\alias{GeoL_readr}
\alias{EaCO_readr}
\alias{MABC_readr}
\alias{FSCA_readr}
\alias{SMC_readr}
\alias{Bal_readr}
\alias{Lic_readr}
\alias{Dor_readr}
\alias{MEPB_readr}
\alias{BBD_readr}
\alias{MPUJ_readr}
\title{Reads specific data files into Darwin Core format}
\usage{
EPEL_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

ASP_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

BMin_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

BMont_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

Ecd_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

Gai_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

KP_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

INHS_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

EcoS_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

GeoL_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

EaCO_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

MABC_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

FSCA_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

SMC_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

Bal_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

Lic_readr(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  beeFilter = TRUE
)

Dor_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

MEPB_readr(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = NULL
)

BBD_readr(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

MPUJ_readr(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = sheet
)
}
\arguments{
\item{path}{The path to the directory containing the data.}

\item{inFile}{The name of the file itself (can also be the remainder of a path including the file name).}

\item{outFile}{The name of the Darwin Core format file to produce.}

\item{dataLicense}{The license to accompany each record in the Darwin Core 'license' column.}
}
\value{
A data frame that's in Darwin Core format.
}
\description{
Data files are specific to various data providers and users may examine code function-by-function.
}
\section{Functions}{
\itemize{
\item \code{ASP_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{BMin_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{BMont_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{Ecd_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{Gai_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{KP_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{INHS_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{EcoS_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{GeoL_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{EaCO_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{MABC_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{FSCA_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{SMC_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{Bal_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{Lic_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{Dor_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{MEPB_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{BBD_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

\item \code{MPUJ_readr()}: Data files are specific to various data providers and users may examine code function-by-function.

}}
\examples{
\dontrun{
An example using a .csv file
EPEL_Data <- EPEL_readr(path = paste0(DataPath, "/Additional_Datasets"),
inFile = "/InputDatasets/bee_data_canada.csv",
outFile = "jbd_EPEL_data.csv",
dataLicense = "https://creativecommons.org/licenses/by-nc-sa/4.0/")

An example using a .xlsx file
    Arm_Data <- Arm_readr(path = paste0(DataPath, "/Additional_Datasets"),
inFile = "/InputDatasets/Bee database Armando_Final.xlsx",
outFile = "jbd_Arm_Data.csv",
sheet = "Sheet1",
dataLicense = "https://creativecommons.org/licenses/by-nc-sa/4.0/")
}
}
