% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{file_finder}
\alias{file_finder}
\title{Finds files within a directory}
\usage{
file_finder(path, fileName)
}
\arguments{
\item{path}{A directory as character. The directory to recursively search.}

\item{fileName}{A character/regex string. The file name to find.}
}
\value{
Returns a directory to the most-recent file that matches the provied fileName. Using regex
can greatly improve specificity. Will also output the file that it has found — it is worthwhile
checking that this is the correct file.
}
\description{
A simple function used to find files within a user-defined directory based on a user-provided character
string.
}
\examples{
file_finder(path = RootPath, fileName = "beesRaw")
# more specifically the .csv version
file_finder(path = RootPath, fileName = "beesRaw.csv")
}
