% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagSummary.R
\name{flagSummary}
\alias{flagSummary}
\title{Build a per-species summary for each and all flags}
\usage{
flagSummary(data = NULL, column = "scientificName", outpath = NULL)
}
\arguments{
\item{data}{A data frame or tibble. The flagged dataset.}

\item{column}{Character. The name of the column to group by and summarise the failed occurrences.
Default = "scientificName".}

\item{outpath}{A character path. Optional. The path and file name in which to save the table as
a .csv. If no outpath is provided, the table will not be saved. Default = NULL.}
}
\value{
A tibble with a column for each flag column (starting with ".") showing the number of
failed (FALSE) occurrences per group. Also shows the (i) total number of records, (ii) total
number of failed records, and (iii) the percentage of failed records.
}
\description{
Takes a flagged dataset and returns the total number of fails (FALSE) per flag (columns starting
with ".") and per species.
Users may define the column to group the summary by. While it is intended to work with
the scientificName column, users may select any grouping column (e.g., country).
}
\examples{
# Load the toy flagged bee data
data("beesFlagged")

  # Run the function and build the flag table
flagTibble <- flagSummary(data = beesFlagged,
                              column = "scientificName",
                              outpath = paste0(tempdir(), "/flagTable.csv"))
                              

}
