% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmoniseR.R
\name{HarmoniseR}
\alias{HarmoniseR}
\title{Harmonise taxonomy of bee occurrence data}
\usage{
HarmoniseR(path = NULL, SynList = BeeDC::beesTaxonomy, occurrences = NULL)
}
\arguments{
\item{path}{A directory as character. The path to a folder that the output can be saved.}

\item{SynList}{A data frame or tibble. The bee taxonomy to use. Default = BeeDC::beesTaxonomy.}

\item{occurrences}{A data frame or tibble. Occurrence records as input.}
}
\value{
The occurrences are returned with update taxonomy columns, including: scientificName,
species, family, subfamily, genus, subgenus, specificEpithet, infraspecificEpithet, and
scientificNameAuthorship. A new column, .invalidName, is also added and is FALSE when the occurrence's
name didn ot match the supplied SynList taxonomy.
}
\description{
Uses the Discover Life taxonomy to harmonise bee occurrences and flag those that do not match
the list. This function could be hijacked to service other taxa if a user matched the format of the
\link{beesTaxonomy} file.
}
\examples{

beesRaw_out <- HarmoniseR(path = DataPath, #The path to a folder that the output can be saved
                       SynList = BeeDC::beesTaxonomy, # The formatted taxonomy file
                       occurrences = BeeDC::beesFlagged)
table(beesRaw_out$.invalidName, useNA = "always")
}
