% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbd_coord_trans.R
\name{jbd_coord_trans}
\alias{jbd_coord_trans}
\title{Internal function. Correct transposed coordinates}
\usage{
jbd_coord_trans(data, x, y, country_code, idcol, worldmap, worldmap_cntr_code)
}
\arguments{
\item{data}{data data.frame. Containing an unique identifier for each
records, geographical coordinates, and country names. Coordinates must be
expressed in decimal degree and in WGS84.}

\item{x}{character string. The column name with longitude data. Default =
"decimalLongitude".}

\item{y}{character string. The column name with latitude data. Default =
"decimalLatitude".}

\item{country_code}{character string. The column name with the country code
assignment of each record. Default = "country_code".}

\item{idcol}{numeric. Dataset unique identifier. Default = "dabase_id".}

\item{worldmap}{polygon. Borders of the world.}

\item{worldmap_cntr_code}{charterer sting. Iso2 code column of country
polygon database.}
}
\value{
Internal function
}
\description{
This functions used different transformations to correct transposed
geographic coordinates.
}
\examples{
\dontrun{
jbd_coord_trans(
  data = occ_country[[i]],
  x = x,
  y = y,
  country_code = cntr_iso2,
  idcol = idcol,
  worldmap = world_poly,
  worldmap_cntr_code = world_poly_iso
  )
}
}
\seealso{
Other prefilter: 
\code{\link{jbd_coordinates_transposed}()}
}
\concept{prefilter}
