% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbd_coordinates_precision.R
\name{jbd_coordinates_precision}
\alias{jbd_coordinates_precision}
\title{Flags coordinates for imprecision}
\usage{
jbd_coordinates_precision(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  ndec = NULL
)
}
\arguments{
\item{data}{A data frame or tibble. Occurrence records as input.}

\item{lat}{Character. The name of the column to use as latitude. Default = "decimalLatitude".}

\item{lon}{Character. The name of the column to use as longitude. Default = "decimalLongitude".}

\item{ndec}{Numeric. The number of decimal places to flag in decimal degrees. Hence, 2 = 0.01.}
}
\value{
Returns the input data frame with a new column, .rou, where FALSE indicates occurrences
that failed the test.
}
\description{
This function flags occurrences where BOTH lat and lon are rounded. This contrasts with
\code{\link[bdc:bdc_coordinates_precision]{bdc::bdc_coordinates_precision()}} that will flag occurrences where only one of lat OR lon are rounded.
The BeeDC approach saves occurrences that may have had terminal zeros rounded.
}
\examples{
beesRaw_out <- jbd_coordinates_precision(
  data = BeeDC::beesRaw,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  ndec = 2 # number of decimals to be tested
)
table(beesRaw_out$.rou, useNA = "always")
}
