% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbd_country_from_coordinates.R
\name{jbd_country_from_coordinates}
\alias{jbd_country_from_coordinates}
\title{Get country names from coordinates}
\usage{
jbd_country_from_coordinates(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country",
  scale = "large"
)
}
\arguments{
\item{data}{A data frame or tibble. This needs to contain, at a minimum,
columns with geographical coordinates and country names.}

\item{lat}{A character string. The column name with latitude in decimal
degrees
and WGS84. Default = "decimalLatitude".}

\item{lon}{A character string. The column with longitude in decimal degrees and
WGS84. Default = "decimalLongitude".}

\item{country}{A character string. The column name with the country assignment
of each record. Default = "country". If no column name is provided a new
column "country" is created.}

\item{scale}{Passed to rnaturalearth's ne_countries().
Scale of map to return, one of 110, 50, 10 or 'small', 'medium', 'large'. Default = "large".}
}
\value{
A tibble of the original input data with country names added for records missing
such information and for which a country determination could be made.
}
\description{
Country names derived from valid geographic coordinates are added to records
missing country names. The prefix, jbd_ is meant
to distinguish this function from the original bdc::bdc_country_from_coordinates().
}
\details{
This function assigns a country name for records missing such
information. Country names are extracted from valid geographic coordinates
using a high-quality map of the world (from the rnaturalearth package). No
country name is added to records whose coordinates are in the sea.
}
\examples{
\donttest{
x <- data.frame(
  decimalLatitude = c(-22.9834, -39.857030, -17.06811, -46.69778),
  decimalLongitude = c(-69.095, -68.443588, 37.438108, -13.82444),
  country = c("", NA, NA, "Brazil"))

bdc_country_from_coordinates(
  data = x,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country"
)
}
}
