% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualOutlierFindeR.R
\name{manualOutlierFindeR}
\alias{manualOutlierFindeR}
\title{Identifies expert outliers and their duplicate records}
\usage{
manualOutlierFindeR(
  occData = NULL,
  DataPath = NULL,
  PaigeOutliersName = "removedBecauseDeterminedOutlier.csv",
  newOutliersName = "All_outliers_ANB.xlsx",
  ColombiaOutliers_all = "All_Colombian_OutlierIDs.csv",
  duplicates = NULL
)
}
\arguments{
\item{occData}{A data frame or tibble. Occurrence records as input.}

\item{DataPath}{A character path to the directory that contains the outlier spreadsheets.}

\item{PaigeOutliersName}{A character file name to the csv file.}

\item{newOutliersName}{A character file name to the xlsx file.}

\item{ColombiaOutliers_all}{A character file name to the csv file.}

\item{duplicates}{A data frame or tibble. The duplicate file produced by \code{\link[=dupeSummary]{dupeSummary()}}.}
}
\value{
Returns the occData with a new column, \link{.expertOutlier} where records that are FALSE are
the expert outliers.
}
\description{
Uses expert-identified outliers with source spreadsheets that may be edited by users. The function
will also use the duplicates file made using \code{\link[=dupeSummary]{dupeSummary()}} to identify duplicates of the
expert outliers and flag those as well.
The function will add a flagging column called \link{.expertOutlier} where records that are FALSE are
the expert outliers.
}
\examples{
\dontrun{
# Read in the most-recent duplicates file as well
if(!exists("duplicates")){
  duplicates <- file_finder(path = DataPath,
                            fileName = "duplicateRun_") \%>\%
    readr::read_csv()}
# identify the outliers and get a list of their database_ids
beesFlagged_out <- manualOutlierFindeR(
  occData = BeeDC::beesFlagged,
  DataPath = DataPath,
  PaigeOutliersName = "removedBecauseDeterminedOutlier.csv",
  newOutliersName = "^All_outliers_ANB_14March.xlsx",
  ColombiaOutliers_all = "All_Colombian_OutlierIDs.csv",
  duplicates = duplicates)
}

}
