% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlagSummary.R
\name{plotFlagSummary}
\alias{plotFlagSummary}
\title{Summary plot of all flags}
\usage{
plotFlagSummary(
  plotData = NULL,
  flagColours = c("#127852", "#A7002D", "#BDBABB"),
  filename = NULL,
  outpath = OutPath_Figures,
  width = 15,
  height = 9,
  units = "in",
  dpi = 300,
  bg = "white",
  device = "pdf",
  speciesName = NULL,
  saveFiltered = FALSE,
  filterColumn = FALSE,
  nameColumn = NULL,
  plotMap = FALSE,
  mapAlpha = 0.5,
  saveTable = FALSE,
  jitterValue = NULL,
  ...
)
}
\arguments{
\item{plotData}{A data frame or tibble. Occurrence records as input.}

\item{flagColours}{A character vector. Colours in order of pass (TRUE), fail (FALSE), and NA.}

\item{filename}{Character. The name of the file to be saved, ending in ".pdf".}

\item{outpath}{Character. The path to the directory in which the figure will be saved.
Default = OutPath_Figures.}

\item{width}{Numeric. The width of the output figure in user-defined units Default = 15.}

\item{height}{Numeric. The height of the output figure in user-defined Default = 9.}

\item{units}{Character. The units for the figure width and height passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
("in", "cm", "mm", or "px").}

\item{dpi}{Passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}. Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or
"screen" (72). Applies only to raster output types. Default = 300.}

\item{bg}{Passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}. Background colour. If NULL, uses the plot.background fill value from the plot theme.
Default = "white."}

\item{device}{Passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}. Device to use. Can either be a device function (e.g. png), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).
Default = "pdf".}

\item{speciesName}{Optional. Character. A species name, as it occurs in the user-input nameColumn.
If entered, the data will be filtered to this species for the plot.
#' @param nameColumn Optional. Character. If speciesName is not NULL, enter the column to look
for the species in. A User might realise that, combined with speciesName, figures can be made for
a variety of factors.}

\item{saveFiltered}{Optional. Logical. If TRUE, the filtered data will be saved to the computer
as a .csv file.
#' @param plotMap Logical. If TRUE, the function will produce a point map. Tested for use with one
species at a time; i.e., with speciesName is not NULL.}

\item{filterColumn}{Optional. The flag column to display on the map. Default = .summary.}

\item{mapAlpha}{Optional. Numeric. The opacity for the points on the map.}

\item{saveTable}{Optional. Logical. If TRUE, the function will save the data used to produce the
compound bar plot.}

\item{jitterValue}{Optional. Numeric. The value to jitter points by in the map in decimal degrees.}

\item{...}{Optional. Extra variables can be fed into \code{\link[forcats:fct_recode]{forcats::fct_recode()}} to change names on plot.
For example... 'B. Mont.' = "BMont", 'B. Minkley' = "BMin", Ecd = "Ecd", Gaiarsa = "Gai"}
}
\value{
Exports a compound bar plot that summarises all flag columns. Optionally can also return
A point map for a particular species in tandem with the summary plot.
}
\description{
Creates a compound bar plot that shows the proportion of records that pass or fail each flag (rows)
and for each data source (columns).
}
\examples{
# import data
data(beesFlagged)
 # Visualise all flags for each dataSource (simplified to the text before the first underscore)
plotFlagSummary(
  plotData = beesFlagged,
  # Colours in order of pass (TRUE), fail (FALSE), and NA
  flagColours = c("#127852", "#A7002D", "#BDBABB"),
  filename = paste0("FlagsPlot_TEST_", Sys.Date(),".pdf"),
  outpath = OutPath_Figures,
  width = 15, height = 9,
  # OPTIONAL:
  #\   #  # Filter to species
  #\   speciesName = "Holcopasites heliopsis",
  #\   # column to look in
  #\   nameColumn = "species",
  #\   # Save the filtered data
  #\   saveFiltered = TRUE,
  #\   # Filter column to display on map
  #\   filterColumn = ".summary",
  #\   plotMap = TRUE,
  #\   # amount to jitter points if desired, e.g. 0.25 or NULL
  #\   jitterValue = NULL,
  #\   # Map opacity value for points between 0 and 1
  #\   mapAlpha = 1,
  # Extra variables can be fed into forcats::fct_recode() to change names on plot
  GBIF = "GBIF", SCAN = "SCAN", iDigBio = "iDigBio", USGS = "USGS", ALA = "ALA", 
  ASP = "ASP", CAES = "CAES", 'B. Mont.' = "BMont", 'B. Minkley' = "BMin", Ecd = "Ecd",
  Gaiarsa = "Gai", EPEL = "EPEL"
)



}
