% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalData_readRs.R
\name{readr_EPEL}
\alias{readr_EPEL}
\alias{readr_ASP}
\alias{readr_BMin}
\alias{readr_BMont}
\alias{readr_Ecd}
\alias{readr_Gai}
\alias{readr_KP}
\alias{readr_EcoS}
\alias{readr_GeoL}
\alias{readr_EaCO}
\alias{readr_MABC}
\alias{readr_FSCA}
\alias{readr_SMC}
\alias{readr_Bal}
\alias{readr_Lic}
\alias{readr_Dor}
\alias{readr_MEPB}
\alias{readr_BBD}
\alias{readr_MPUJ}
\title{Reads specific data files into Darwin Core format}
\usage{
readr_EPEL(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_ASP(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_BMin(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_BMont(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_Ecd(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_Gai(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_KP(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_EcoS(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_GeoL(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_EaCO(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_MABC(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_FSCA(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_SMC(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_Bal(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_Lic(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  beeFilter = TRUE
)

readr_Dor(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_MEPB(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = NULL
)

readr_BBD(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_MPUJ(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = sheet
)
}
\arguments{
\item{path}{The path to the directory containing the data.}

\item{inFile}{The name of the file itself (can also be the remainder of a path including the file name).}

\item{outFile}{The name of the Darwin Core format file to produce.}

\item{dataLicense}{The license to accompany each record in the Darwin Core 'license' column.}
}
\value{
A data frame that's in Darwin Core format.
}
\description{
Data files are specific to various data providers and users may examine code function-by-function.
}
\section{Functions}{
\itemize{
\item \code{readr_ASP()}: Reads specific data files into Darwin Core format

\item \code{readr_BMin()}: Reads specific data files into Darwin Core format

\item \code{readr_BMont()}: Reads specific data files into Darwin Core format

\item \code{readr_Ecd()}: Reads specific data files into Darwin Core format

\item \code{readr_Gai()}: Reads specific data files into Darwin Core format

\item \code{readr_KP()}: Reads specific data files into Darwin Core format

\item \code{readr_EcoS()}: Reads specific data files into Darwin Core format

\item \code{readr_GeoL()}: Reads specific data files into Darwin Core format

\item \code{readr_EaCO()}: Reads specific data files into Darwin Core format

\item \code{readr_MABC()}: Reads specific data files into Darwin Core format

\item \code{readr_FSCA()}: Reads specific data files into Darwin Core format

\item \code{readr_SMC()}: Reads specific data files into Darwin Core format

\item \code{readr_Bal()}: Reads specific data files into Darwin Core format

\item \code{readr_Lic()}: Reads specific data files into Darwin Core format

\item \code{readr_Dor()}: Reads specific data files into Darwin Core format

\item \code{readr_MEPB()}: Reads specific data files into Darwin Core format

\item \code{readr_BBD()}: Reads specific data files into Darwin Core format

\item \code{readr_MPUJ()}: Reads specific data files into Darwin Core format

}}
\examples{
\dontrun{
An example using a .csv file
EPEL_Data <- Ereadr_PEL(path = paste0(DataPath, "/Additional_Datasets"),
inFile = "/InputDatasets/bee_data_canada.csv",
outFile = "jbd_EPEL_data.csv",
dataLicense = "https://creativecommons.org/licenses/by-nc-sa/4.0/")

An example using a .xlsx file
    Arm_Data <- readr_Arm(path = paste0(DataPath, "/Additional_Datasets"),
inFile = "/InputDatasets/Bee database Armando_Final.xlsx",
outFile = "jbd_Arm_Data.csv",
sheet = "Sheet1",
dataLicense = "https://creativecommons.org/licenses/by-nc-sa/4.0/")
}
}
