% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoFinder.R
\name{repoFinder}
\alias{repoFinder}
\title{Find GBIF, ALA, iDigBio, and SCAN files in a directory}
\usage{
repoFinder(path)
}
\arguments{
\item{path}{A directory as character. The path within which to recursively look for GBIF, ALA,
iDigBio, and SCAN files.}
}
\value{
Returns a list of directories to each of the above data downloads
}
\description{
Find GBIF, ALA, iDigBio, and SCAN files in a directory
}
\examples{
\dontrun{
# Where DataPath is made by [BeeBDC::dirMaker()]
BeeBDC::repoFinder(path = DataPath)
}
}
