% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USGS_formatter.R
\name{USGS_formatter}
\alias{USGS_formatter}
\title{Find, import, and format USGS data to Darwin Core}
\usage{
USGS_formatter(path, pubDate)
}
\arguments{
\item{path}{A character path to a directory that contains the USGS data, which will be found using
\code{\link[=fileFinder]{fileFinder()}}. The function will look for "USGS_DRO_flat".}

\item{pubDate}{Character. The publication date of the dataset to update the metadata and citation.}
}
\value{
Returns a list with the occurrence data, "USGS_data", and the EML data, "EML_attributes".
}
\description{
The function finds, imports, formats, and creates metadata for the USGS dataset.
}
\examples{
\dontrun{
USGS_data <- USGS_formatter(path = DataPath, pubDate = "19-11-2022")
}

}
