% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beesChecklist.R
\docType{data}
\name{beesChecklist}
\alias{beesChecklist}
\title{A country-level checklist of bees from Discover Life}
\format{
An object of class \code{"tibble"}
\describe{
\item{validName}{The valid scientificName as it should occur in the scientificName column.}
\item{ISO}{The ISO names as it occurs on Discover Life.}
\item{Long_country_name}{The country names as it occurs on Discover Life.}
\item{Country_name}{A short country name version.}
\item{Alpha-2}{Alpha-2 from rnaturalearth.}
\item{Alpha-3}{Alpha-3 from rnaturalearth.}
\item{ISO.y}{entry}
\item{database_id}{Another format of Alpha-3.}
\item{official}{Official country name "yes" or Discover Life name "no"}
\item{rNaturalEarth_name}{Country name from rnaturalearth's name_long.}
\item{Country_or_island}{Country or island name from Discover Life.}
\item{Notes}{Discover Life country name notes.}
\item{Freq}{The count of unique species names per country.}
\item{matchName}{A simplified validName/scientificName for matching.}
\item{accid}{The accepted taxon id according to beesTaxonomy. For accepted taxa, accid == 0.}
\item{id}{The id number for the name in beesTaxonomy.}
\item{family}{Bee family.}
\item{subfamily}{Bee subfamily.}
\item{validName_taxonomy}{The validName from the beesTaxonomy}
\item{canonical}{The validName without the scientificNameAuthority.}
\item{canonical_withFlags}{The validName without the scientificNameAuthority but with Discover Life flags.}
\item{genus}{Bee genus.}
\item{subgenus}{Bee subgenus.}
\item{species}{Bee specificEpithet.}
\item{infraspecies}{Bee infraSpecificEpithet.}
\item{authorship}{Bee scientificNameAuthorship.}
\item{taxon_rank}{Rank of the taxon name.}
\item{matchCertainty}{Quality of the name match to the Discover Life checklist.}
\item{sciName_country}{A simplified merger of validName and country for matching.}

}
}
\usage{
BeeDC::beesChecklist
}
\description{
Contains taxonomic and country information for the bees of the world. See \code{\link[=beesTaxonomy]{beesTaxonomy()}}
for further context
}
\examples{

data(beesChecklist)
head(beesChecklist)

}
\references{
This dataset was created using the Discover Life checklist and taxonomy. See rawDataWrite.R
for further information on download date in file name.
Dataset is from the publication:
DOREY, J. B., CHESSHIRE, P. R., BOLAÑOS, A. N., O’REILLY, R. L., BOSSERT, S., COLLINS, S. M., LICHTENBERG, E. M., TUCKER, E., SMITH-PARDO, A., FALCON-BRINDIS, A., GUEVARA, D. A., RIBEIRO, B. R., DE PEDRO, D., FISCHER, E., HUNG, J. K.-L., PARYS, K. A., ROGAN, M. S., MINCKLEY, R. L., VELZCO, S. J. E., GRISWOLD, T., ZARRILLO, T. A., SICA, Y., ORR, M. C., GUZMAN, L. M., ASCHER, J., HUGHES, A. C. & COBB, N. S. In review. A globally synthesised and flagged bee occurrence dataset and cleaning workflow. Scientific Data.
}
\keyword{datasets}
