% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryOutliers.R
\name{countryOutlieRs}
\alias{countryOutlieRs}
\title{Flag country-level outliers with a provided checklist.}
\usage{
countryOutlieRs(
  checklist = NULL,
  occData = NULL,
  keepAdjacentCountry = TRUE,
  pointBuffer = NULL,
  rnearthScale = 50
)
}
\arguments{
\item{checklist}{A data frame or tibble. The formatted checklist built based on the Discover Life website.}

\item{occData}{A data frame or tibble. The a Darwin Core occurrence dataset.}

\item{keepAdjacentCountry}{Logical. If TRUE, occurrences in countries that are adjacent to checklist countries will be
kept. If FALSE, they will be flagged.}

\item{pointBuffer}{Numeric. The buffer around points beyond which if they are outside of a
checklist country they will be flagged as FALSE. This provides a good way to maintian points that
occur right along the coast of the rnaturalearth maps.}

\item{rnearthScale}{Numeric. The map scale fed into \code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}}'s scale parameter:
Scale of map to return, one of 110, 50, 10 or 'small', 'medium', 'large', where smaller numbers
are higher resolution. WARNING: This function is tested on 110 and 50.}
}
\value{
The input data with a new column, .countryOutlier. TRUE == passed, FALSE == failed,
NA == did not overlap with rnaturalearth map.
}
\description{
Flag country-level outliers with a provided checklist.
}
\examples{

beesRaw_out <- countryOutlieRs(checklist = BeeDC::beesChecklist,
                               occData = BeeDC::beesRaw,
                               keepAdjacentCountry = TRUE,
                               pointBuffer = 0.05,
                               rnearthScale = 50)
table(beesRaw_out$.countryOutlier, useNA = "always")
}
