% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonAlley.R
\name{diagonAlley}
\alias{diagonAlley}
\title{Find fill-down errors}
\usage{
diagonAlley(data = NULL, minRepeats = NULL)
}
\arguments{
\item{data}{A data frame or tibble. Occurrence records as input.}

\item{minRepeats}{Numeric. The minimum number of lat or lon repeats to flag a record}
}
\value{
Retrusn the data with a new column, .sequential, where FALSE = records that have consecutive
latitudes or longitudes.
}
\description{
A simple function that looks for potential lat/lon fill down errors by finding consecutive numbers.
}
\examples{
# Read in the example data
  data(beesRaw)
 # Run the function
  beesRaw_out <- diagonAlley(
    data = beesRaw,
    # The minimum number of repeats needed to find a sequence in for flagging
    minRepeats = 4)
  

}
