% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmoniseR.R
\name{harmoniseR}
\alias{harmoniseR}
\title{Harmonise taxonomy of bee occurrence data}
\usage{
harmoniseR(
  data = NULL,
  path = NULL,
  taxonomy = BeeBDC::beesTaxonomy,
  speciesColumn = "scientificName",
  rm_names_clean = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble. Occurrence records as input.}

\item{path}{A directory as character. The path to a folder that the output can be saved.}

\item{taxonomy}{A data frame or tibble. The bee taxonomy to use. Default = BeeBDC::beesTaxonomy.}

\item{speciesColumn}{Character. The name of the column containing species names. Default = "scientificName".}

\item{rm_names_clean}{Logical. If TRUE then the names_clean column will be removed at the end of
this function to help reduce confusion about this column later. Default = TRUE}
}
\value{
The occurrences are returned with update taxonomy columns, including: scientificName,
species, family, subfamily, genus, subgenus, specificEpithet, infraspecificEpithet, and
scientificNameAuthorship. A new column, .invalidName, is also added and is FALSE when the occurrence's
name did not match the supplied taxonomy.
}
\description{
Uses the Discover Life taxonomy to harmonise bee occurrences and flag those that do not match
the checklist. This function could be hijacked to service other taxa if a user matched the format of the
\link{beesTaxonomy} file.
}
\examples{

beesRaw_out <- harmoniseR(
  #The path to a folder that the output can be saved
path = tempdir(),
# The formatted taxonomy file
taxonomy = BeeBDC::beesTaxonomy, 
data = BeeBDC::beesFlagged,
speciesColumn = "scientificName")
table(beesRaw_out$.invalidName, useNA = "always")
}
