% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOccurrences.R
\name{importOccurrences}
\alias{importOccurrences}
\title{Imports the most-recent repoMerge data}
\usage{
importOccurrences(path = path, fileName = "^BeeData_")
}
\arguments{
\item{path}{A directory as a character. The directory to recursively look in for the above data.}

\item{fileName}{Character. A String of text to look for the most-recent dataset.
Default = "^BeeData_". Find faults by modifying \code{\link[=fileFinder]{fileFinder()}}
and logic-checking the file that's found.}
}
\value{
A list with a data frame of merged occurrence records, "Data_WebDL", and a list of EML
files contained in "eml_files".
}
\description{
Looks for and imports the most-recent version of the occurrence data created by the \code{\link[=repoMerge]{repoMerge()}}
function.
}
\examples{
\dontrun{
DataImp <- importOccurrences(path = DataPath)
}
}
