% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbd_coordCountryInconsistent.R
\name{jbd_coordCountryInconsistent}
\alias{jbd_coordCountryInconsistent}
\title{Flags coordinates that are inconsistent with the stated country name}
\usage{
jbd_coordCountryInconsistent(
  data = NULL,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  mapResolution = 50,
  pointBuffer = 0.01
)
}
\arguments{
\item{data}{A data frame or tibble. Occurrence records as input.}

\item{lon}{Character. The name of the column to use as longitude. Default = "decimalLongitude".}

\item{lat}{Character. The name of the column to use as latitude. Default = "decimalLatitude".}

\item{mapResolution}{Numeric or character. To be passed to \code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}}'s scale.
Scale of map to return, one of 110, 50, 10 or “small”, “medium”, “large”.
Smaller values return higher-resolution maps.}

\item{pointBuffer}{Numeric. Amount to buffer points, in decimal degrees. If the point is outside
of a country, but within this point buffer, it will not be flagged.}
}
\value{
The input occurrence data with a new column, .coordinates_country_inconsistent
}
\description{
Compares stated country name in an occurrence record with record’s coordinates using
rnaturalearth data. The prefix, jbd_ is meant
to distinguish this function from the original bdc::bdc_coordinates_country_inconsistent().
}
\examples{

beesRaw_out <- jbd_coordCountryInconsistent(
  data = BeeBDC::beesRaw,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  mapResolution = 50,
  pointBuffer = 0.01)
}
