% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbd_Ctrans_chunker.R
\name{jbd_Ctrans_chunker}
\alias{jbd_Ctrans_chunker}
\title{Wraps jbd_coordinates_transposed to identify  and fix transposed occurrences}
\usage{
jbd_Ctrans_chunker(
  data = NULL,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  id = "databse_id",
  country = "country_suggested",
  countryCode = "countryCode",
  sci_names = "scientificName",
  border_buffer = 0.2,
  save_outputs = TRUE,
  stepSize = 1e+06,
  chunkStart = 1,
  append = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble. Occurrence records as input.}

\item{lat}{Character. The name of the column to use as latitude. Default = "decimalLatitude".}

\item{lon}{Character. The name of the column to use as longitude. Default = "decimalLongitude".}

\item{id}{Character. Default = "database_id".}

\item{country}{Character. The name of the column containing country names. Default = "country'.}

\item{countryCode}{Character. Identifies the countryCode column. Default = "countryCode".}

\item{sci_names}{Character. The columns containing scientific names. Default = "scientificName".}

\item{border_buffer}{Numeric. The buffer, in decimal degrees, around points to help match them
to countries. Default = 0.2 | ~22 km at equator.}

\item{save_outputs}{Logical. If TRUE, transposed occurrences will be saved to their own file.}

\item{stepSize}{Numeric. The number of occurrences to process in each chunk. Default = 1000000.}

\item{chunkStart}{Numeric. The chunk number to start from. This can be > 1 when you need to restart
the function from a certain chunk; for example if R failed unexpectedly.}

\item{append}{Logical. If TRUE, the function will look to append an existing file.}
}
\value{
Returns the input data frame with a new column, coordinates_transposed, where FALSE = columns
that had coordinates transposed.
}
\description{
Because the \code{\link[=jbd_coordinates_transposed]{jbd_coordinates_transposed()}} function is very RAM-intensive, this wrapper
allows a user to specify chunk-sizes and only analyse a small portion of the occurrence data at a
time. The prefix jbd_ is used to highlight the difference between that function and the related
\code{\link[bdc:bdc_coordinates_transposed]{bdc::bdc_coordinates_transposed()}}.
}
\examples{
beesFlagged_out <- jbd_Ctrans_chunker(
# bdc_coordinates_transposed inputs
data = BeeDC::beesFlagged,
id = "database_id",
lat = "decimalLatitude",
lon = "decimalLongitude",
country = "country_suggested",
countryCode = "countryCode",
border_buffer = 0.2, # in decimal degrees (~22 km at the equator)
save_outputs = TRUE,
sci_names = "scientificName",
# chunker inputs
stepSize = 1000000,  # How many rows to process at a time
chunkStart = 1,  # Start row
append = FALSE  # If FALSE it may overwrite existing dataset
) 
table(beesFlagged_out$coordinates_transposed, useNA = "always")

}
