% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalData_readRs.R
\name{readr_CAES}
\alias{readr_CAES}
\alias{readr_MABC}
\alias{readr_Col}
\alias{readr_Bal}
\alias{readr_Arm}
\alias{readr_MEPB}
\alias{readr_MPUJ}
\alias{readr_JoLa}
\title{Reads specific data files into Darwin Core format}
\usage{
readr_CAES(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = "Sheet1"
)

readr_MABC(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = "Hoja1"
)

readr_Col(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = sheet
)

readr_Bal(path = NULL, inFile = NULL, outFile = NULL, dataLicense = NULL)

readr_Arm(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = "Sheet1"
)

readr_MEPB(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = NULL
)

readr_MPUJ(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = sheet
)

readr_JoLa(
  path = NULL,
  inFile = NULL,
  outFile = NULL,
  dataLicense = NULL,
  sheet = c("pre-1950", "post-1950")
)
}
\arguments{
\item{path}{A character path. The path to the directory containing the data.}

\item{inFile}{Character or character path. The name of the file itself (can also be the
remainder of a path including the file name).}

\item{outFile}{Character or character path. The name of the Darwin Core format file to be saved.}

\item{dataLicense}{Character. The license to accompany each record in the Darwin Core 'license' column.}

\item{sheet}{A character String. For those datasets read from an .xlsx format, provide the
sheet name.}
}
\value{
A data frame that is in Darwin Core format.
}
\description{
Data files are specific to various data providers and users may examine code function-by-function.
}
\section{Functions}{
\itemize{
\item \code{readr_MABC()}: Reads specific data files into Darwin Core format

\item \code{readr_Col()}: Reads specific data files into Darwin Core format

\item \code{readr_Bal()}: Reads specific data files into Darwin Core format

\item \code{readr_Arm()}: Reads specific data files into Darwin Core format

\item \code{readr_MEPB()}: Reads specific data files into Darwin Core format

\item \code{readr_MPUJ()}: Reads specific data files into Darwin Core format

\item \code{readr_JoLa()}: Reads specific data files into Darwin Core format

}}
\examples{
\dontrun{
# An example using a .xlsx file
    Arm_Data <- readr_Arm(path = paste0(tempdir(), "/Additional_Datasets"),
inFile = "/InputDatasets/Bee database Armando_Final.xlsx",
outFile = "jbd_Arm_Data.csv",
sheet = "Sheet1",
dataLicense = "https://creativecommons.org/licenses/by-nc-sa/4.0/")
}
}
