% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMaps.R
\name{summaryMaps}
\alias{summaryMaps}
\title{Create country-level summary maps of species and occurrence numbers}
\usage{
summaryMaps(
  mapData = NULL,
  class_n = 15,
  class_Style = "fisher",
  filename = NULL,
  width = 10,
  height = 5,
  dpi = 300
)
}
\arguments{
\item{mapData}{A data frame or tibble. Occurrence records as input.}

\item{class_n}{The number of categories to break the data into.}

\item{class_Style}{The class style passed to \code{\link[classInt:classIntervals]{classInt::classIntervals()}}, options are chosen style: one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks", "dpih", "headtails", or "maximum".
Default = fisher}

\item{filename}{A character vector with both the path to the save location and the file name itself
for the output figure.}

\item{width}{Numeric. The width, in inches, of the resulting figure. Default = 10.}

\item{height}{Numeric. The height, in inches, of the resulting figure. Default = 5.}

\item{dpi}{Numeric. The resolution of the resulting plot. Default = 300.}
}
\value{
Saves a figure to the user-specified outpath and name with a global map of bee occurrence
species and count data from the input dataset.
}
\description{
Builds an output figure that shows the number of species and the number of occurrences per country.
Breaks the data into classes for visualisation. Users may filter mapData to their taxa of interest to produce
figures of interest.
}
\examples{
# This simple example using the test data has very few classes due to the small amount of input data.
summaryMaps(
mapData = BeeDC::beesFlagged,
width = 10, height = 10,
class_n = 4,
class_Style = "fisher",
filename = paste0(OutPath_Figures, "/CountryMaps_fisher_TEST.pdf")
)


}
